+++ hd.h
@@ -45,7 +45,7 @@ struct console_info {
 int build_initial_search_string(char * search_path, char * search_string);
 //create_footer(char *footer_string, short console_width, char *root_path, char search_drive);
 char * create_horizontal_line(char *, CONSOLE_SCREEN_BUFFER_INFO);
-char * compact_size_with_suffix(int, char *);
+char * compact_size_with_suffix(long long, char *);
 int display_footer();
 int display_header(char * search_path);
 int display_help();
